fix: align ProjectID type to uint32 across OpenAPI and provider (#6)

+++ internal/services/project/project_resource.go
@@ -446,7 +446,7 @@ func boolFromPtr(v *bool) types.Bool {
 	return types.BoolValue(*v)
 }
 
-func parseProjectID(s string) (int64, error) {
+func parseProjectID(s string) (uint32, error) {
 	return client.ParseProjectID(s)
 }
